Update last_word.c
changed the condition while ( I > 0) to while ( i >= 0) because if there was only one word and nothing else in the string ( like for example "test") it would print the word but leave out its first letter. The reason behind this issue is the fact that in our example , "t" is str[0], but zero is excluded due to the condition.